backend/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-3-jet==1.0.8
django-cors-headers==3.5.0
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.5.0
packaging==20.4
Pillow==7.2.0
pre-commit==2.7.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
zipp==3.1.0
Faker==4.14.0
